# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_server_handlers.ClaimctlReservation:
    properties:
      createdAt:
        $ref: "#/definitions/pgtype.Int8"
      duration:
        $ref: "#/definitions/pgtype.Interval"
      endTime:
        $ref: "#/definitions/pgtype.Int8"
      id:
        type: integer
      queuePosition:
        $ref: "#/definitions/pgtype.Int4"
      resourceId:
        type: integer
      scheduledEndTime:
        $ref: "#/definitions/pgtype.Timestamptz"
      startTime:
        $ref: "#/definitions/pgtype.Int8"
      status:
        $ref: "#/definitions/pgtype.Text"
      updatedAt:
        $ref: "#/definitions/pgtype.Int8"
      userId:
        type: integer
    type: object
...
  pgtype.InfinityModifier:
    enum:
      - 1
      - 0
      - -1
    format: int32
    type: integer
    x-enum-varnames:
      - Infinity
      - Finite
      - NegativeInfinity
  pgtype.Int4:
    properties:
      int32:
        format: int32
        type: integer
      valid:
        type: boolean
    type: object
  pgtype.Int8:
    properties:
      int64:
        format: int64
        type: integer
      valid:
        type: boolean
    type: object
  pgtype.Interval:
    properties:
      days:
        format: int32
        type: integer
      microseconds:
        format: int64
        type: integer
      months:
        format: int32
        type: integer
      valid:
        type: boolean
    type: object
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: "#/definitions/pgtype.InfinityModifier"
      time:
        type: string
      valid:
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@github.com/thetaqitahmid/claimctl
    name: API Support
  description: API documentation for claimctl resource management system.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: claimctl API
  version: "1.0"
paths:
...
  /reservations:
    post:
      consumes:
        - application/json
      description: Create a new reservation for a resource.
      parameters:
        - description: Reservation Request
          in: body
          name: request
          required: true
          schema:
            properties:
              resourceId:
                type: integer
            type: object
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/internal_server_handlers.ClaimctlReservation"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a reservation
      tags:
        - reservations
  /reservations/{id}/activate:
    patch:
      consumes:
        - application/json
      description:
        Activate a pending reservation. Only accessible by admin or resource
        owner.
      parameters:
        - description: Reservation ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/internal_server_handlers.ClaimctlReservation"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Activate a reservation
      tags:
        - reservations
  /reservations/{id}/cancel:
    patch:
      consumes:
        - application/json
      description: Cancel a pending or active reservation.
      parameters:
        - description: Reservation ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cancel a reservation
      tags:
        - reservations
  /reservations/{id}/complete:
    patch:
      consumes:
        - application/json
      description:
        Complete an active reservation. Only accessible by reservation
        owner or admin.
      parameters:
        - description: Reservation ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Complete a reservation
      tags:
        - reservations
  /reservations/timed:
    post:
      consumes:
        - application/json
      description: Create a new reservation for a resource with a specific duration.
      parameters:
        - description: Timed Reservation Request
          in: body
          name: request
          required: true
          schema:
            properties:
              duration:
                type: string
              resourceId:
                type: integer
            type: object
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/internal_server_handlers.ClaimctlReservation"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a timed reservation
      tags:
        - reservations
...
swagger: "2.0"