# backend/internal/server/handlers/doc_types.go
package handlers

import "github.com/thetaqitahmid/claimctl/internal/db"

// Aliases for Swagger documentation
type ClaimctlResource = db.ClaimctlResource
type ClaimctlReservation = db.ClaimctlReservation
type ClaimctlUser = db.ClaimctlUser